break out AppBookmarkField
